refactor: move app js to stimulus controllers
