+++ app/views/home/index.html.erb
@@ -1,6 +1,6 @@
 <%- if current_user && current_user.credentials.any? %>
-  <div id="add-credential">
-    <%= form_with scope: :credential, url: credentials_path, remote: true do |form| %>
+  <div data-controller="add-credential">
+    <%= form_with scope: :credential, url: credentials_path, remote: true, data: { action: "ajax:success->add-credential#create" } do |form| %>
       <div class="mdc-text-field mdc-text-field--box" data-controller="textfield">
         <%= form.text_field :nickname, class: "mdc-text-field__input"%>
         <label class="mdc-floating-label" for="my-text-field">Nickname</label>
